+++ R/dic.R
@@ -50,3 +50,73 @@ xDIC = function(temp, pH, S) {
   data.frame(xCO2, xHCO3, xCO3)
 
 }
+
+##———————————————————————————————————————————————————————————————————————————##
+#### xabs ####
+#' @title Relative rates of CO2 absorption reactions
+#'
+#' @description `xabs()` calculates the relative abundance of the DIC species
+#'   as a function of solution temperature, pH, and salinity.
+#'
+#' @param temp The temperature of the solution (°C).
+#' @param pH The pH of the solution.
+#' @param S The salinity of the solution (g/kg or ‰).
+#'
+#' @details
+#'
+#' xhydration = ((kCO2 / (kCO2 + kOHxKw / aH)) * 100), where
+#'
+#' * kCO2 is the rate constant for CO2 hydration from Johnson (1982)
+#' * kOHxKw is the rate constant for
+#'   CO2 hydroxylation x Kw from Schulz et al. (2006).
+#' * aH is 10^(-pH)
+#'
+#' @return
+#' Returns a data frame with the relative rates of CO2 absorption reactions:
+#' * Relative rate of CO2 hydration (%).
+#' * Relative rate of CO2 hydroxylation (%).
+#'
+#' @references
+#' Johnson, K. S. (1982).
+#' Carbon dioxide hydration and dehydration kinetics in seawater.
+#' Limnology and Oceanography, 27(5), 894-855.
+#' <https://doi.org/10.4319/lo.1982.27.5.0849>
+#'
+#' Schulz, K. G., Riebesell, U., Rost, B., Thoms, S., & Zeebe, R. E. (2006).
+#' Determination of the rate constants for the carbon dioxide to
+#' bicarbonate inter-conversion in pH-buffered seawater systems.
+#' Marine Chemistry, 100(1-2), 53-65.
+#' <https://doi.org/10.1016/j.marchem.2005.11.001>
+#'
+#' @examples
+#' xabs(temp = 25, pH = 7, S = 30)
+#'
+#' @export
+
+xabs = function(temp, pH, S) {
+  TinK = temp  + 273.15
+  aH = 10 ^ -pH
+
+  # Rate constant for CO2 (aq) hydration Johnson (1982)
+  A_CO2 = 1246.98
+  B_CO2 = 0
+  D_CO2 = -6.19 * 10 ^ 4
+  E_CO2 = -183.0
+  kCO2 = exp(A_CO2 + B_CO2 * sqrt(S) + D_CO2 / TinK + E_CO2 * log(TinK))
+
+  # Rate constant for CO2 (aq) hydroxylation x Kw from Johnson (1982)
+  A_OH = -930.13
+  B_OH = 0.110
+  D_OH = 3.10 * 10 ^ 4
+  E_OH = 140.9
+  kOHxKw = exp(A_OH + B_OH * sqrt(S) + D_OH / TinK + E_OH * log(TinK))
+
+  # Rate constant for CO2 (aq) hydroxylation x Kw from Schulz et al. (2006)
+  kOHxKw = (499002.24 * exp(4.2986 * 10 ^ -4 * S ^ 2 + 5.75499 * 10 ^
+                              -5 * S)) * exp(-90166.83 / (8.3145 * TinK))
+
+  xhydration     = ((kCO2 / (kCO2 + kOHxKw / aH)) * 100)
+  xhydroxylation = 100 - ((kCO2 / (kCO2 + kOHxKw / aH)) * 100)
+
+  final = return(data.frame(xhydration, xhydroxylation))
+}