+++ tests/testthat/test_clumped.R
@@ -1,29 +1,23 @@
-#### tests of the clumped isotope functions ####
+#### test the clumped functions ####
+
+
+# ----- Test warnings -----
 
 test_that("D47 and D48 error if eq is not or wrongly specified", {
-  expect_error(D47(10))
   expect_error(D47(10, "cheese"), "Invalid input for eq")
-  expect_error(D48(10))
   expect_error(D48(10, "cheese"), "Invalid input for eq")
 })
-test_that("D47 and D48 error if temp is not or wrongly specified", {
-  expect_error(D47("cheese", "Fiebig21"))
-  expect_error(D48("cheese", "Fiebig21"))
+
+test_that("temp_D47 if eq is not or wrongly specified", {
+  expect_error(temp_D47(0.600, eq = "cheese"), "Invalid input for eq")
 })
 
-# Additional tests for all valid input parameters
+
+# ----- Test outputs -----
+
 test_that("D47 and D48 return values", {
   expect_type(D47(10, eq = "Petersen19"), "double")
   expect_type(D48(10, eq = "Swart21"), "double")
-})
-
-
-test_that("temp_D47 and temp_D48 error if clumped error and eq is not or wrongly specified", {
-  expect_error(temp_D47(0.600))
-  expect_error(temp_D47("grated", "cheese"))
-  expect_error(temp_D47(0.600, eq = "cheese"), "Invalid input for eq")
-  expect_error(temp_D48("cheese", 0.139, 0.002, 0.010, ks = -0.6))
-  expect_error(temp_D48(0.600))
 })
 
 test_that("temp_D48 returns a single value if error IS NOT specified", {
@@ -41,6 +35,10 @@ test_that("temp_D48 returns three values as data frame if error IS specified", {
   expect_length(temp_D47(0.617, 0.005, eq = "Petersen19"), 2)
   expect_type(temp_D47(0.617, 0.005, eq = "Petersen19"), "double")
 })
-test_that("temp_D48 error if values are wrongly specified", {
-  expect_error(temp_D48(temp_D48("cheese", 0.139, 0.002, 0.010, ks = -0.6)))
-})
+
+# ----- Test graphics -----
+
+plot(1:10)
+expect_message(temp_D48(0.617, 0.139, 0.002, 0.010, ks = -0.6, add = T), "Graphics added to the plot")
+graphics.off()
+