+++ tests/testthat/test_utils.R
@@ -1,6 +1,11 @@
 # tests of the utils #
 
 
+test_that("utility functions error if eq is not or wrongly specified", {
+  expect_error(to_VSMOW(10, "cheese"), "Invalid input for eq")
+  expect_error(to_VPDB(10, "cheese"), "Invalid input for eq")
+})
+
 test_that("utility functions return a number", {
   expect_type(to_VSMOW(0), "double")
   expect_type(to_VPDB(0), "double")
@@ -11,9 +16,6 @@ test_that("utility functions return a number", {
   expect_type(a_A_B(10, 12), "double")
   expect_type(A_from_a(10, 1.01), "double")
   expect_type(B_from_a(10, 1.01), "double")
+  expect_type(epsilon(1.01), "double")
 })
 
-test_that("utility functions error if eq is not or wrongly specified", {
-  expect_error(to_VSMOW(10, "cheese"), "Invalid input for eq")
-  expect_error(to_VPDB(10, "cheese"), "Invalid input for eq")
-})